Добавлен модуль руководителя

+++ WorkingDiary/StartupWindow.xaml.cs
@@ -36,7 +36,7 @@ namespace WorkingDiary
             if (logInWindow.ShowDialog() == true)
             {
                
-                int id = GetWorkerId(logInWindow.Login, logInWindow.Password);
+                long id = GetWorkerId(logInWindow.Login, logInWindow.Password);
                 using (WorkersDbContext db = new())
                 {
                     checkWorker = db.Workers.Find(id);
@@ -84,12 +84,12 @@ namespace WorkingDiary
 
             infoWindow.Show();
         }
-        int GetWorkerId(string login, string password)
+        long GetWorkerId(string login, string password)
          {
              using (WorkersDbContext workersDb = new())
                  return workersDb.Workers.Where(w => w.WorkerLogin.Equals(login) && w.WorkerPassword.Equals(password)).Select(w => w.WorkerId).FirstOrDefault();
          }
-        Worker GetWorker(int id)
+        Worker GetWorker(long id)
          {
              using (WorkersDbContext workersDb = new())
              {
